+++ game_patch/misc/alpine_settings.cpp
@@ -365,6 +365,10 @@ bool alpine_player_settings_load(rf::Player* player)
         g_alpine_game_config.fps_counter = std::stoi(settings["ShowFPS"]);
         processed_keys.insert("ShowFPS");
     }
+    if (settings.count("ShowSpeed")) {
+        g_alpine_game_config.speed_display = std::stoi(settings["ShowSpeed"]);
+        processed_keys.insert("ShowSpeed");
+    }
     if (settings.count("FPSCounterAverageMs")) {
         g_alpine_game_config.set_fps_counter_average_ms(std::stoi(settings["FPSCounterAverageMs"]));
         processed_keys.insert("FPSCounterAverageMs");
@@ -1100,6 +1104,7 @@ void alpine_player_settings_save(rf::Player* player)
     file << "BurningEntityLights=" << g_alpine_game_config.burning_entity_lights << "\n";
     file << "SpectateMinimalUI=" << g_alpine_game_config.spectate_mode_minimal_ui << "\n";
     file << "ShowFPS=" << g_alpine_game_config.fps_counter << "\n";
+    file << "ShowSpeed=" << g_alpine_game_config.speed_display << "\n";
     file << "FPSCounterAverageMs=" << g_alpine_game_config.fps_counter_average_ms << "\n";
     file << "SaveConsoleHistory=" << g_alpine_game_config.save_console_history << "\n";
     file << "AlpineBranding=" << g_alpine_game_config.af_branding << "\n";